backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20

# Milvus and OpenAI
pymilvus>=2.4.0
openai>=1.35.0

# Document processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.3.0

# Additional dependencies
numpy>=1.26.0
pydantic>=2.11.0
pydantic-settings>=2.0.0
typing-extensions>=4.14.0
